Add files via upload
This is a screen recording of an example scene rendered in a 3D graphics engine I implemented. A (textured) grass landscape sits above (a textured plane of) water, with a ball with a man's face texture mapped onto it floats above the landscape. More interestingly, a lighting and shadowing have also been implemented. The user has control over the camera's rotation, the camera's position, the camera's "zoom", and the position of the light (but not the rotation of the light).

(Note, the landscape generating code and the 3D mesh data structure were written by Professor Josh Davis.)